fix variable manager

+++ src/backEnd/Managers/VariableManager.java
@@ -31,6 +31,7 @@ public class VariableManager {
     }
 
     public void setValueToVariable(String v, double value){
+    	System.out.println("setValueToVariable: "+v+" "+value);
     	String variable = deleteVariableSyntax(v);
     	myVariableMap.put(variable, value);
     }
@@ -39,12 +40,26 @@ public class VariableManager {
      * Returns the variable name by deleting its Variable syntax (e.g. :)
      */
     private String deleteVariableSyntax(String v){
-    	if(v.contains(VARIABLE_PROGRAM_SYNTAX))
-    	return v.substring(myProgramLanguage.getString(VARIABLE_PROGRAM_SYNTAX).length());
+    	System.out.println("deleteVariableSyntax called: "+v);
+    	System.out.println("substring index: "+myProgramLanguage.getString(VARIABLE_PROGRAM_SYNTAX).length());
+    	System.out.println("v length: "+v.length() +" "+v);
+    	
+    	if(v.contains(myProgramLanguage.getString(VARIABLE_PROGRAM_SYNTAX))){
+        	v = v.substring(myProgramLanguage.getString(VARIABLE_PROGRAM_SYNTAX).length());	
+    	}
+    	System.out.println("deleteVariableSyntax Done: "+v);
     	return v;
     }
     
     public void removeVariable(String v){
+    	System.out.println("removeVariable: "+v);
+    	if(v.startsWith(myProgramLanguage.getString(VARIABLE_PROGRAM_SYNTAX))){
+        	v = deleteVariableSyntax(v);
+    	}
+    	for(String s: myVariableMap.keySet()){
+    		System.out.print("key: "+s+" ");
+    		System.out.println();
+    	}
     	myVariableMap.remove(v);
     }
     
@@ -53,9 +68,13 @@ public class VariableManager {
      * If v is not initialized, return null
      */
     public double getValueOfVariable(String v){
-    	String variable = deleteVariableSyntax(v);
-    	return myVariableMap.get(variable);
+    	if(v.startsWith(myProgramLanguage.getString(VARIABLE_PROGRAM_SYNTAX))){
+        	String variable = deleteVariableSyntax(v);	
+        	return myVariableMap.get(variable);
+    	}
+    	return myVariableMap.get(v);
     }
+    
     public void readFromFile(File file) throws FileNotFoundException{
     	myVariableMap=new HashMap<String,Double>();
     	try {
@@ -70,6 +89,7 @@ public class VariableManager {
 			throw new FileNotFoundException();
 		}
     }
+    
     public String mapToString(String separatingCharacter){
     	String result="";
     	for(String s:myVariableMap.keySet()){
